Run elf-cleaner in multiple threads if multiple input are given
When we want to loop over many files we can then save quite some time
by running on parallel inside termux-elf-cleaner, instead of outside
in the shell.

Basically, changing from:

  parallel termux-elf-cleaner foo bar baz foo bar baz ...

and instead handle the parallelisation inside termux-elf-cleaner so we
can run just:

  termux-elf-cleaner foo bar baz foo bar baz ...

decreases the amount of overhead and saves time compared to running
termux-elf-cleaner in a bash for loop, or with the help of tools like
parallel.